requirements.txt

streamlit==1.39.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3